addaaddaadasdfasdf as

+++ WebApplication_14/src/java/dao/UserDAO.java
@@ -1,9 +1,7 @@
 /*
-DAO là một design pattern dùng để truy cập và thao tác với database
-Chứa tất cả các phương thức CRUD (Create, Read, Update, Delete)
-Đóng gói toàn bộ logic truy cập database
-Tách biệt logic truy cập dữ liệu khỏi business logic
-Giúp code dễ maintain và test hơn
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
  */
 package dao;
 
@@ -21,17 +19,17 @@ import utils.DBUtils;
 
 /**
  *
- * @author tungi
+ * @author tungiF
  */
 public class UserDAO implements IDAO<UserDTO, String> {
 
     @Override
     public boolean create(UserDTO entity) {
         String sql = "INSERT [dbo].[tblUsers] ([userID], [fullName], [roleID], [password]) "
-                + "VALUES (?, ?, ?, ?)";
-
+                + "VALUES (?, ? ,? ,?)";
+        Connection conn;
         try {
-            Connection conn = DBUtils.getConnection();
+            conn = DBUtils.getConnection();
             PreparedStatement ps = conn.prepareStatement(sql);
             ps.setString(1, entity.getUserID());
             ps.setString(2, entity.getFullName());
@@ -50,11 +48,12 @@ public class UserDAO implements IDAO<UserDTO, String> {
     @Override
     public List<UserDTO> readAll() {
         List<UserDTO> list = new ArrayList<>();
-        String sql = "SELECT * FROM [tblUsers];";
+        String sql = "SELECT * FROM [tblUsers]";
         try {
             Connection conn = DBUtils.getConnection();
+            conn = DBUtils.getConnection();
             PreparedStatement ps = conn.prepareStatement(sql);
-            ResultSet rs = ps.executeQuery();
+            ResultSet rs = ps.executeQuery(sql);
             while (rs.next()) {
                 UserDTO user = new UserDTO(
                         rs.getString("userID"),
@@ -73,7 +72,7 @@ public class UserDAO implements IDAO<UserDTO, String> {
 
     @Override
     public UserDTO readById(String id) {
-        String sql = "SELECT * FROM tblUsers WHERE userID = ?";
+        String sql = "SELECT * FROM tblUsers WHERE userID= ?";
         try {
             Connection conn = DBUtils.getConnection();
             PreparedStatement ps = conn.prepareStatement(sql);
@@ -100,11 +99,11 @@ public class UserDAO implements IDAO<UserDTO, String> {
         String sql = "UPDATE [tblUsers] SET "
                 + "[fullName] = ?, "
                 + "[roleID] = ?, "
-                + "[password] = ? "
+                + "[password] =? "
                 + "WHERE [userID] = ?";
-
+        Connection conn;
         try {
-            Connection conn = DBUtils.getConnection();
+            conn = DBUtils.getConnection();
             PreparedStatement ps = conn.prepareStatement(sql);
             ps.setString(1, entity.getFullName());
             ps.setString(2, entity.getRoleID());
@@ -122,13 +121,13 @@ public class UserDAO implements IDAO<UserDTO, String> {
 
     @Override
     public boolean delete(String id) {
-        String sql = "DELETE FROM [tblUsers] "
-                + "WHERE [userID] = ?";
+        String sql = "DELETE FROM [tblUsers] WHERE [userID] = ?";
+        Connection conn;
         try {
-            Connection conn = DBUtils.getConnection();
+            conn = DBUtils.getConnection();
             PreparedStatement ps = conn.prepareStatement(sql);
             ps.setString(1, id);
-            int n = ps.executeUpdate(sql);
+            int n = ps.executeUpdate();
             return n > 0;
         } catch (ClassNotFoundException ex) {
             Logger.getLogger(UserDAO.class.getName()).log(Level.SEVERE, null, ex);
